requirements.txt
opencv-python
numpy
matplotlib
pandas
opencv-contrib-python
mediapipe
scikit-learn
python-dotenv